+++ app/Http/Controllers/GondolaController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Model\Gondola;
+use App\Http\Controllers\Search;
+use App\Model\Product;
 
 class GondolaController extends Controller{
 
@@ -57,10 +59,14 @@ class GondolaController extends Controller{
 
     public function delete($id = null){
         $this->gondola = Gondola::find($id);
-        if($this->gondola->delete())
-            return redirect()->route('gondola.list')->with('success', "Gôndola Excluida com Sucesso.");
-        else
-            return redirect()->back('gondola.show', $id)->with('error', "Erro ao Excluir Gôndola.");
+
+        if(count(Product::where('gondola_id', $this->gondola->id)->get()) == 0){
+            if($this->gondola->delete())
+                return redirect()->route('gondola.list')->with('success', "Gôndola Excluida com Sucesso.");
+            else
+                return redirect()->back('gondola.show', $id)->with('error', "Erro ao Excluir Gôndola.");
+        }else
+            return redirect()->back()->with('excluir', 'Gôndola Está Vinculada a um Produto.');
     }
 
     public function search(Request $request){
@@ -72,22 +78,7 @@ class GondolaController extends Controller{
         if(is_Null($request->description) && is_Null($request->criado) && is_Null($request->atualizado))
             $models = Gondola::orderBy('id', 'asc')->paginate(5);
         
-        if(isset($request->description))
-            $model = Gondola::where('description', 'like', "%$request->description%");//orderBy('id', 'asc')->paginate(5);
-            
-        if(isset($request->criado)){
-            $created_at = explode(' - ', $request->criado);
-            if(isset($model))
-                $model->whereDate('created_at','>=', $created_at[0])->whereDate('created_at','<=', $created_at[1]);
-            else
-                $model = Gondola::whereDate('created_at','>=', $created_at[0])->whereDate('created_at','<=', $created_at[1]);
-        }if(isset($request->atualizado)){
-            $updated_at = explode(' - ', $request->atualizado);
-            if(isset($model))
-                $model->whereDate('updated_at','>=', $updated_at[0])->whereDate('updated_at','<=', $updated_at[1]);
-            else
-                $model = Gondola::whereDate('updated_at','>=', $updated_at[0])->whereDate('updated_at','<=', $updated_at[1]);
-        }
+        $model = Search::quest($request->description, $request->criado, $request->atualizado, Gondola::class);
 
         $models = is_Null($models) ? $model->orderBy('id', 'asc')->paginate(5) : $models;
